Allow the syntax "import module"
This is done by locating all uses of the module in the file and replacing module.func with module_func using redbaron. Once this is done the remainder can be traversed as before. The line being read is converted to the syntax "from module import func" and this phrase is also traversed as before