Mark values which are tested for "is None" as optional
This is done by replacing the ast.core.Variable with an ast.core.ValuedVariable
The ast.core.ValuedVariable is created using the ast.core.Variable.clone function.
This function is adapted to allow modification of properties.

This fixes the fortran translation (and therefore the problem in pyccel). But it does not fix epyccel as f2py passes 0 in place of optional variables. A wrapper must therefore be added