Test "if a is None" in pyccel instead of epyccel
Currently f2py does not support optional arguments with a default value of None (https://github.com/numpy/numpy/issues/4013).
The code can however be tested in a pure fortran implementation.
The tests in the epyccel folder have been kept and marked as xfail, however it is unlikely that these will become xpass as long as f2py is used as the relevant issue is now 6 years old.