Select correct source for semantic stage
If objects are imported from a program then choose either the program file or the module file for each target.
It is likely that this will not compile. Either because we cannot import objects from a program or because the module name is wrong. The module name is generated with a random aspect. This will only match if the parser which represents this file is the same that is used for printing. Alternatively it will generate valid fortran code with only a small modification required for the compilation to succeed. A warning is therefore raised in parser/parser.py