Print function calls returning multiple variables on their own line
This is done for the functions : print, min, max
There may be places that are missing. A warning is therefore raised if _print_Function is called with multiple outputs. This function is not used by _print_Assign and should therefore not need to be used if there are multiple outputs.