Multiple function outputs (#262)
* Allow multiple outputs in f2py interface

* Keep ordering of function results

* Test functions with multiple results

* Allow assignment of multiple outputs in fortran

This is necessary if functions return more than one basic type.
It is achieved by creating a function _visit_lhs_Assign in parser/semantic.py
This function is called for the lhs variable or multiple times if this variable is a tuple

* [ERRORS] Improve output number mismatch error

Use pyccel error system instead of raising an error

* Add names to output arguments of subroutines

* Collect result for _visit_lhs_Assign

* Correct tuple assignment

* Correct fortran printing if there are out arguments but no input arguments

* Add function call tests for functions with multiple arguments

TODO: Fix function call in print or raise error

* Add test_multiple_results to test_pyccel. Ensure all outputs are correctly tested.

* Print function calls returning multiple variables on their own line

This is done for the functions : print, min, max
There may be places that are missing. A warning is therefore raised if _print_Function is called with multiple outputs. This function is not used by _print_Assign and should therefore not need to be used if there are multiple outputs.

* Correct variable name

* Remove debugging prints

* Get function without raising an error to check results. This is necessary for interfaces and macros

* Check against correct class type

* Add missing output types. Strip python output to handle tuples

* Allow for tuples printed like lists with []

* Use regexes to parse fortran and python output

* Add tests for min/max. Add scientific notation to regex

* Simplify method for adding additional code before a line

Create variable _additional_code to be accessed when parsing the body of a function or program
This variable is set in _visit_Function for functions with multiple outputs

* Use CodeBlock for Program body

* Use CodeBlock for While body

* Use CodeBlock for For body. Remove _iprint (_print_Block no longer returns 2 results)

* Use CodeBlock for For body

* Use CodeBlock for FunctionDef body

* Use CodeBlock for Block, inline and With.

* Fix IfTernaryOperator

* Fix complexity

* Add tests for variables with multiple outputs modified in a CodeBlock

* Remove unnecessary list conversion

* Add IfTernaryOperator test

* Rename _visit_lhs_Assign function as _assign_lhs_variable

Co-authored-by: Emily.Bourne <louise.bourne@gmail.com>