Fix generator function variable duplication errors
A variable is temporarily created as an int so that the generator functions can find the variable named
A new variable is then created with the correct type. This leads to incorrect errors being raised (see issue #272).
The tests have additional problems as the temporary variable was not removed. This push removes that temporary variable