Finish fixes (relies on merge from PR 334)
Correctly identify all objects that are pointers
Prevent pointers for homogeneous tuples. This is a fortran restriction. In fortran we can create a pointer to an array but not an array of pointers.
TODO: See if the fortran restriction can be moved to the codegen stage. Perhaps by modifying _print_Alias_Assign and _print_Indexed