Move fortran restrictions to codegen/fcode.py.
This will allow us to use arrays of pointers in other langauges (e.g. C).
TupleVariables are examined when entering a Program Module or FunctionDef and all Tuples which are non homogeneous in Fortran are marked as such.
This requires special handling in _print_IndexedElement in order to return the element variable and in AliasAssign to assign elements individually.
Other places (e.g. _print_Assign) already handled not is_homogeneous as a special case