Move obsolete files to new folder (#343)
The following files:

    pyccel/commands/ipyccel.py
    pyccel/commands/build.py
    pyccel/commands/quickstart.py
    pyccel/codegen/cmake.py
    pyccel/codegen/templates/*
    pyccel/calculus/finite_differences.py

have 0% coverage. This has led to them not being correctly maintained. Some have useful code but which does not belong in pyccel. Some have code which we may wish to reintegrate into pyccel once their contents have been updated to be compatible with pyccel in its current form.

All of these files have been moved to a new folder obsolete/ which will not be installed and will not be checked when calculating coverage.