Correct tuple behaviour to correctly use pointers (#336)
* Add tuples tests to epyccel to increase granularity
* Add test for issue 328
* Split test to also check homogeneous case
* Add pointer test for copy
* Remove pyccel tuples tests which have been moved to epyccel
* Ensure that the tuple used for the homogeneous test is actually homogeneous
* Add teardown to functions in epyccel tests
* Remove setup/teardown for each function on parallel tests
* Ensure that there is only one instance of SystemRandom (at module level)

This reduces the probability to generate modules with same name when calling epyccel on functions. Note that SystemRandom may not be available on all operating systems (https://docs.python.org/3/library/random.html#random.SystemRandom), but we choose it because it uses urandom, a system-level random number generator with cryptographic quality.

* Finish fixes (relies on merge from PR 334)

Correctly identify all objects that are pointers.
Prevent pointers for homogeneous tuples (fortran restriction: cannot create an array of pointers).

* Move fortran restrictions to codegen/fcode.py.

This will allow us to use arrays of pointers in other languages (e.g. C).
TupleVariables are examined when entering a Program Module or FunctionDef and all Tuples which are non homogeneous in Fortran are marked as such.
This requires special handling in _print_IndexedElement in order to return the element variable and in AliasAssign to assign elements individually.
Other places (e.g. _print_Assign) already handled not is_homogeneous as a special case

* Allow precision=None for strings (fixes issue #338)
* Rely on previously set pointer information rather than duplicating code
* Type and only use l.is_pointer if l has this property
* Add back in removed code to correctly handle a = (b,c)
* Print alias assign between 2 TupleVariables correctly
* Update version