Correct string printing
Apostrophes are not used for quoting strings in every language (e.g. in C apostrophes are used for characters while quotes are used for strings)
In addition the character used for quoting must be escaped differently in different languages.
Thus the quoting must be done in the codegen step not the syntactic step