Store imports under alias
If no targets are specified then the imported module can have an alias. The code will only be aware of this alias. The lookup will therefore be done via the alias name.
However the original name must be conserved to allow the imports to be written correctly in Fortran.
The AsName class is modified so that it is hashed by target name. This allows it to be used as a key which will be recognised if the alias is provided:
i.e:
   > a = {AsName('numpy','np'):1, AsName('scipy','sp'):2}
        {numpy as np: 1, scipy as sp: 2}
   > 'np' in a
        True
   > 'numpy' in a
        False