Simplify _print_Bool
The third argument of `merge` is a boolean. It is therefore redundant to do:
`merge(.true., .false., bool)`

Parentheses will be added in the syntactic stage depending on the precedence of any operators used in the expression. Writing them explicitly here may lead to two sets of parentheses and will lead to unnecessary parentheses (e.g. `(x) /= 0` )