Fix broken codegen tests (#437)
Ensure that codegen tests fail if there is a problem during their treatment, even if the problem creates a neat error (fixes #436). Fix broken tests. Improve `broadcast` function by supposing that mathematical expressions are valid. As a result when both shapes are unknown in a dimension they are supposed to be the same, when only one shape is unknown then it is supposed that it is the same as the known shape (fixes #245).

- Stop codegen tests passing when errors have been raised
- Print filename with error (useful for errors in imported files)
- Allow broadcast function to handle PyccelArraySize
- Correct handling of elemental functions
- Add end to end test for elemental
- Do not complain about dtype differences in an AugAssign
- Assume that unknown sizes on the same axis are the same size
- When one of the sizes is unknown, prefer the known size