Use correct precision when printing integers (#446)
Use the precision property to print the correct precision. Casting functions are used where possible so this modification only needs to be done in one place. References to sympy object `IndexedBase` are removed.
A .codacy.yml file is also added to stop checking the tests folder for duplication. Duplication here is normal

- Use IndexedVariable instead of sympy's IndexedBase
- Print integer with correct precision. Fixes #326
- In ast/builtins.py, rename `Bool` -> `PythonBool` and `Int` -> `PythonInt`
- Use casting function in PyccelDiv (Fortran & C) and PyccelMod (Fortran)
- Use casting functions for NumpySqrt (Fortran) and MathSqrt (Fortran)
- Use casting functions where possible for PyccelFloorDiv (Fortran), provide precision elsewhere
- Use default precision for casting functions in C
- Remove print functions for Sympy objects
- [CODACY] Do not check for duplication in tests
- [COVERAGE] Add test to ensure allocatables are not used in an expression