Add non-MPI Windows tests with GitHub Actions (#558)
Add GitHub Actions for Windows with Python 3.7.

Unfortunately Python 3.8 leads to linking problems because `setup-python` installs the most recent patch, and the new version of Python 3.8 collides with the local version whos dll has already been identified (e.g Python 3.8.1 is uninstalled and Python 3.8.3 replaces it).

The use of `setup-python` means we do not use `conda`. As a result we are missing a windows patch in distutils. This is a common problem for windows users as detailed here:
https://stackoverflow.com/questions/34135280/valueerror-unknown-ms-compiler-version-1900
The install step `Fix 'Unknown MS Compiler version 1900' problem` handles this problem. It may not be necessary if Python was installed with Anaconda.

TODO:
- Fix MPI installation