Ensure literals have the right argument. Create literals with a requested precision (#572)
Add type checks to Literal classes to ensure that they contain arguments of the correct type. Use the precision argument to translate precision modifiers of literals to a literal of the correct precision. E.g, `np.float32(2)` is now translated to `2.0_C_FLOAT` instead of `a = Real(2_C_LONG_LONG, 4)`. Casts are still printed for truncated values. Commit summary:

- Use inheritance to set `Literal` precision and check its type
- Remove unused `sympy.Integer` inheritance from `LiteralInteger` (this has many knock-on effects as it makes it obvious when we are not using pyccel operators)
- Add type checks and conversions to `Literal` classes
- Use literal precision to translate expressions such as `np.float32(2)` directly to `2.0_C_FLOAT`
- Print floats with correct precision in fortran
- Fix `unexpected keyword argument : "precision"`
- ensure that `complex` function works when complex arguments are passed
- Simplify phrases such as `2+5j` to `LiteralComplex(2,5)`