Allow translating from Python to Python (#652)
As explained in issue #643 the Python printer is no longer up to date. Even basic things such as `return` do not work. This is because this code is never tested. In order to avoid problems in Psydac we should therefore test translation to Python too. This PR makes translating from Python to Python possible and activates some of the few tests which work.

**Commit Summary**

- Update Python Import printer;
- When translating to Python stop after code generation and copy file to output location;
- Stop pyccel command from overwriting original Python file;
- Only check for shared library when not translating to Python;
- Activate some tests;
- Group python tests in the workflow.