Remove IndexedVariable type (#647)
An `IndexedElement` has 2 properties, `base` and `indexes`: for `a[i]`, `a`  is the base and `i`  is the index.

Previously the base was an `IndexedVariable` (most of the time). An `IndexedVariable` has most of the properties of a `Variable` but none of the newer ones (e.g allow_negative_indexes) as it was not well maintained. This means that the `Variable` has to be accessed via `a.internal_variable`. This leads to the code becoming less readable without providing any benefits. This PR removes the unnecessary `IndexedVariable` type which means that the base of an `IndexedElement` will always be up to date with the `Variable` type.

For the behaviour to remain the same in the rest of the code the magic `__getitem__` method is implemented for `Variable` and updated for `TupleVariable`. Previously the `__getitem__` method of a `TupleVariable` was only used when providing 1 index to a non-homogeneous tuple. It therefore returned the internal `Variable` at that index. It must now handle the indexing for homogeneous tuples (which is the same as for a normal `Variable`) and for inhomogeneous tuples (the first index is treated as before, the additional indices are passed to the `__getitem__` method of the returned `Variable`).

**Commit Summary**

- Remove `IndexedVariable` class;
- Make `IndexedElement` base into `Variable`;
- Add `__getitem__` method to `Variable`;
- Update `__getitem__` method of `TupleVariable`;
- Remove references to `IndexedVariable`;
- Remove unnecessary code at end of `_extract_indexed_from_var` function;
-  Add empty slices in `IndexedElement` constructor instead of `SemanticParser`;
- Simplify `IndexedElement` constructor;
- Remove obsolete methods `_print_IndexedBase` and `_print_Indexed` from Fortran printer.