Remove sympy.Tuple from pyccel's AST representation (#676)
This PR removes `sympy.Tuple` from Pyccel's AST representation, because it is not necessary. Where it is used to group items a python `tuple` is sufficient. Where it may represent tuples in the translated code `PythonTuple` should be used instead. Unless `sympy.Tuple` is created with `sympify = False` it also converts its arguments to sympy objects, making it a common source of errors. This fixes #657.