Make classes in ast/core.py mutable (#713)
As mentioned in issue #660, most classes use the _args variable (or .args property) provided by the sympy inheritance to store their arguments. This means that the arguments are immutable and several of our properties are hard to maintain. This PR makes all classes in ast/core.py mutable, as a result the majority of `__new__` functions are removed from this file.

`Argument` no longer inherits from `Symbol`. As a result the pyccel class `Lambda` had to be created to replace the sympy class `Lambda` (which only accepts `Symbol` arguments).