Add cast to Literal and tuple
Python objects are often given as arguments. This should not happen for objects which can be children, but adding Literal calls all over the code is costly and hard to remember. It is simpler to correct this in the initialisation

In order to simplify isinstance calls, it is also better if all iterable children have the same type. Currently some are lists, most are tuples, occasionally they are dict_values or other unexpected objects. To harmonise this, all iterables are saved as tuples