Access templates of a function via its decorators (#683)
Currently, we have two elements (`templates` which is a dictionary of `Template` objects, and `template` which is a list of decorator calls in the `decorators` dictionary) in `FunctionDef` class that have infos in the same theme (templates). As templates are not an integral part of a function, it was decided to group this information in the `decorators` dictionary (see #591).

This PR adds a `template` key to the `decorators` dictionary. This item contains the `template` list (of decorators) and the `templates` dictionary. As a result we don't need to pass templates dictionary as an argument to `FunctionDef` since it will be accessible from `decorators` dictionary through the `template` key. It therefore fixes #591.