Handle type casting in array_fill (#724)
`array_fill` is a Pyccel builtin function, written in C, which is used to fill an `array`(second argument) with a given `value` (first argument) and return a `value-filled array`. The type of `value` should be the same type of `array` to prevent wrong alignments and even crashes (in case of the data type of `value` greater than the data type of `array`). By fixing we can reactivate the x-failing C tests for `NumpyFull` and `NumpyFullLike` (soon), and fix all kinds of casting in `array_fill`.

Commit summary:
- Handle the casting (including complex casting) in `NumpyFull` class by using more flexible dictionaries and function to handle them;
- Cast dynamically in `_print_PythonComplex`, `_print_PythonFloat`, `_print_PythonInt`;
- Activate the test `test_full_dtype` in test_numpy_functions.py`.