Ensure one-line statements in printers end with `\n (#830)
Make sure that all one-line statements are printed with a final \n as done in Fortran. This prevents unnecessary empty lines being printed, as \n does not need to be added later when joining lines. This fixes #761. The updated printers are:

- CCodePrinter
- PythonPrinter
- CWrapperCodePrinter

In addition multiple pylint errors are fixed in the python printer (unused import, typo in variable, wildcard import).