Improve selection of compiler & linker for Windows and Intel (#863)
Fix #859:
- Remove need for system-wide compiler choice on Windows
- Compile and link wrapper with Intel when compiling generated sources with Intel

On Windows provide the argument to the `setup_X.py` file when the compiler is from the intel or gcc families, to avoid the need for the hacky modification of the `distutils.cfg` file.  Provide 2 possible names for the gfortran library on Windows. Set environment variables in the correct place to correctly compile and link files with Intel.