Fix 1D-1D matmul bug (#875)
This pull request fixes 1D-1D matmul bug that raises `ValueError: Incompatible rank in variable allocation`. It does this by changing the rank to 0 and the shape to an empty tuple, since the result will be a scalar. It also corrects the printing for this case. The tests for that are activated now.

In addition fix #853 (wrong dtype for matmul of booleans) by adding `NativeBool` as possible `dtype` of `NumpyMatmul`.