Allow stand-alone expressions (#902)
Allow a stand-alone expression, which appears as a statement by itself with its return value not used or stored. Notably, this handles the case of a function call whose results are not collected (fixes #896).

This is achieved by creating a temporary variable to collect the result of any stand-alone expression. The variable is then dropped if the expression is a function call with no return. This is contrary to the method used in PR #852 where useless statements were removed, but allows us to correctly handle expressions such as `f()+g()`

**Commit Summary**

- Create a temporary variable to collect the result of any stand-alone expression (fixes #896);
- Remove warning created by PR #852 which can no longer be raised;
- Ensure a `TupleVariable` can be created from the results of a function call;
- Add tests;
- Correct `is_temp` collection in `create_new_variable`;
- Move warning test to codegen.