Add transpose function (#939)
Implement the Numpy `transpose` function. In Fortran when used as a temporary variable the `transpose` function is used. When the variable is saved, the ordering is used to ensure pointers are created. In C the function is handled through unravelling or with a new function `transpose_alias_assign`.

The `transpose` function is also used to implement the ndarray property `.T` and method `.transpose()`.

Related to #806 and #458.