Add shortcuts to Fortran literals (#941)
Close issue #856. Add dictionary `iso_c_binding_shortcut` with shortcuts, to be used in Fortran instead of the original `iso_c_binding` dictionary whenever possible, so that e.g. `4_C_LONG_LONG` can be printed as `4_i64`. The same shortcuts are used in the variables' declarations.

Checking if shortcuts can be implemented is done during the initialization of `FCodePrinter` through a systematic check of all the variable names in the namespace, whose result is then stored in the attribute `_namespace_restrictions` of the aforementioned class as an array of booleans of size 8.

The renaming is done in the `print_kind` method whenever it is possible. Finally, a line is added to state all the macros used in the module.