Add dummy variables where necessary to avoid aliasing in Fortran  (#1013)
If a variable is passed to a function call as argument, and is also used for storing the result (`intent(out)`), use a copy (dummy variable) as argument in the generated Fortran code. This prevents the out variable from being wiped before it is used as the argument. Fixes #981.

BUGFIX: Don't check returns inside sub-functions when checking return compatibility.