Improve imports to allow importing inline functions (#1046)
Improve import infrastructure so that all imports are aware of what they are importing (rather than just knowing its name). This allows us to identify whether an imported object is an inline function. The import of a module is modified from being described as a dictionary to being described by a `Module` object. When printing imports import all necessary objects.

The inline functions are now exposed to Python. As the wrappers call inline functions, the contents of the function are inlined into the wrapper.

When header files are added in the future (see issue #297) inline functions should be printed into the header file so they are available for inlining.

**Commit Summary**

- [BUGFIX] Use Variables in `BindCFunctionDef` body not `FunctionDefArgument`s
- Ensure that the renamed object inside `AsName` is a pyccel object
- Add an internal dictionary to `Module` type to allow searching inside a `Module`
- Reorganise `Iterable` to remove the need for importing all possible future iterables (potentially causing circular imports)
- Add `global_funcs` to `InlineFunctionDef`
- Create `PyccelFunctionDef` class
- Ensure that `Import`s are aware of the object being imported after the syntactic stage
- Create `Decorator` class
- Save external libraries into `Module` instances
- Make `ast.numpyext.Shape` into a class so it can be stored in `PyccelFunctionDef`
- Move scipy information to its own file
- Add comparison magic methods to `DottedName`
- Import any objects necessary for inline methods
- Remove unused function `CWrapperCode._print_IndexedElement`
- Wrap inline functions
- [BUGFIX] Not all Python `ast` object have line and column information
- [BUGFIX] In `get_filename_from_import` we were importing the Python translation of a module, if it was available, rather than the original module because the current working directory is inside `__pyccel__`
- Save functions in a header file into its `Module`
- Remove duplicate functions in `stdlib.internal.lapack`. Keep the correct function signature
- Add a test for imports
- Remove restriction from documentation