Clean up scope: Part 1 (#1068)
In order to facilitate future development on issues such as #227, it is important to have a clean Scope class that is easy to work with from anywhere in pyccel. This is the first PR cleaning up the scope. This PR does not change the way pyccel works at all. It just moves the Scope to its own file and puts in place some of the functions which will be useful later. This will make future PRs easier to follow.

**Commit Summary**
- Move `Scope` class to its own file
- Change storage system inside `Scope` to avoid forgetting objects inside the `imports`
- Add accessor functions to `Scope` to interact with `Scope` directly rather than via the `SemanticParser`
- Remove static functions
- Remove `@python` decorator. Fixes #668