Improve printing for Fortran (#953)
This PR aims to improve the printing capabilities of Fortran.

The printing of all expressions in Fortran was done through the use of `', '.join(['print *', *args])`. This approach is limited in the following ways:

- Padding was added automatically in many cases and this hinders readability.
- A new line is printed even if not sought.
- All arrays are printed the same way regardless of their rank.
- Although formatting can be achieved with print, it was not used.

All these issues were addressed with the use of `write` whenever it is possible. Expressions are processed as they would in C to generate a format for the arguments. this format will be passed to `write` for an output as close as possible to C and Python. In addition, arrays with rank > 1 are printed as they would in C.

Solves #982 (Can't print empty string).
Solves #940 (Printing nd arrays in Fortran).

Co-authored-by: EmilyBourne <louise.bourne@gmail.com>