Add naming module (#1089)
Add naming modules to define rules for when there are clashes (e.g. with language specific keywords, with variables beginning with `_`, case sensitive). Use the naming module inside the Scope to ensure that symbols do not clash. Map the Python names to the final names and take care to use the appropriate name in each part of the code (final name except when searching through headers). Pass a `PyccelSymbol` to `insert_symbol` when the object is not generated by Pyccel.

Rename forbidden variables which appear in OpenMP statements (which are not parsed explicitly - see #590).

Delete `tests/pyccel/scripts/hope_benchmarks_decorators/hope_fib.py` which is no longer needed as we can now translate `tests/pyccel/scripts/hope_benchmarks_decorators/fib.py`.

The most complicated part of this PR handles the module name. In the case where the module and a function share the same name the choice was made to change the module name as Fortran is the only language where there can be collisions between the name of a module and a function. However, as the file is not renamed this means that the printing stage needs to be aware of the old and the new name to import correctly in languages that use a module name (i.e. Fortran) and languages that use a filename (C, Python). This is handled through an `AsName` object. Fixes #227.

**Additional Commits**

- Add `__setstate__` methods to `Scope` and `BasicParser` to ensure that incompatible pickle files are correctly regenerated
- Fix `DeprecationWarnings` from pytest concerning `pyccel/tests/ndarrays/conftest.py`
- Remove unused properties
- Set blocker status of errors so that all errors are blocking in developer-mode