Fix slow pyccel behaviour (#1113)
When calling `get_attribute_nodes` or `substitute` we must always exclude `FunctionDef` unless we are searching for a `FunctionDef`. Otherwise Pyccel will search through all the contents of any functions called within the context. This is the cause of the bug in issue #1114 .

This PR fixes #1114 by redefining the class method `ignore` for `FunctionClass` so it never recurses into a `FunctionDef`. Redundant `excluded_nodes=(FunctionDef,)` commands are removed.