Benchmarks (#1035)
Add all code necessary to benchmark Pyccel's performance. Comparing it to Python, Pythran, Numba, and PyPy (if available and requested). Intel vs GCC is also compared upon request.

Pushes to the master branch trigger an update of the `performance.md` file which shows the performance results. This operation takes 2h so must only be run on the master branch, but it is possible to make it 3 times faster by reducing the accuracy and not reporting error bounds (using time instead of pyperf).