Fix bug preventing exporting arrays with an uppercase letter in the name (#1155)
Fortran is case-insensitive, hence all symbols are saved in the .o file as lower case values. When calling a Fortran function from C, which is case sensitive, everything needs to be lower case. This was not done in function `wrap_module_array_var` of our `bind_c` files, which led to linker errors.

**List of changes**

- Bug-fix: ensure that `bind_c` names only contain lower case letters
- Add a unit test