Fix race condition (#1158)
On some systems, clock problems can cause the stdlib folders to always be copied. In a multithread environment this might lead to a thread trying to compile with those files while another thread replaces them (by deleting the folder and copying the contents) see #1092  and https://github.com/pyccel/psydac/issues/227. To fix this, this PR adds locks for the files inside the folder which match the locks used by the compiler. Fixes #1092 . Fixes https://github.com/pyccel/psydac/issues/227

The call to `getmtime` is also replaced by a call to `getatime` which prevents the unnecessary copies on some systems.