Set order to None for non array variable (#836)
The `order` attribute of non-array variables must be set to `None`. Failure to do so was a bug, which caused "inhomogeneous arguments" errors in the case of tuples or lists with variables and literals of the same type. Fixes issue #835.

Further, change the default value of `shape` for scalars from `()` to `None`.

Co-authored-by: EmilyBourne <louise.bourne@gmail.com>