Fix the precision of numpy.abs function (#1177)
Fix the precision of `numpy.absolute` function and the internal precision of the real and imaginary parts of a `LiteralComplex`. The latter caused no bugs so far as the precision of the complex was used to interpret the value and no literals with precisions higher than the default precision can be translated. Fixes #1176