Ensure cuda files are installed with pyccel (#1345)
Include cuda files in the manifest to ensure that they are installed
with pyccel. Fixes #1344 . This should help with the coverage as the
python files that are run are the installed versions not the local
versions.

Use coverage's `[paths]` options to rename files tested on different
systems. Fixes #1346 .

Pull some updates from master regarding the installation and test setup