Add a bot to Pyccel to improve PRs (#1352)
Add a bot to Pyccel to make PRs more friendly and automatise some
housekeeping tasks. The bot does the following:

- Adds a welcome message to a new PR
- Keeps PRs in draft until ready
- Adds messages to encourage reviews
- Ensure branches have the correct review label

Additionally the tests are organised so that all tests can be run in the
review process, including tests only used on the master branch, and
tests on other Python versions. The doc is also updated. A CHANGELOG is
added which should now be updated with 1 line for each pull request.