Enforce pylint conventions (#1359)
Add an additional test to complain when unexpected pylint disable
commands are added. This test recommends the use of pylint disable
commands for docstrings in the tests. Additionally the disabling of
docstrings for `_print_X` and `_visit_X` functions is now handled in the
.pylintrc file. This means that this only applies to those functions
instead of applying to all the functions in the file.