Improve coverage check (#1369)
Ensure that coverage is not compulsory for the bot to mark the PR as
ready. This is done by asking the user to explicitly tell the bot to
ignore failing coverage tests. In the same message the user is also
asked to explain the failures and open appropriate issues. Fixes #1368.
Additionally, discard code in `__str__` or `__repr__` functions from
coverage output. These functions are used for debugging so the testing
requirements do not need to be as stringent.

The bot is also patched so the flags are removed when it knocks a PR
back to draft due to failing tests.