Fix linking error with Anaconda (#1367)
Link Python library using its absolute path. Fixes #1366.

Previously the Python library was linked without specifying the path
explicitly. This is the standard and most flexible way of linking
libraries, but it caused problems in an environment using Anaconda as it
forces the inclusion of the Anaconda paths in the library directories.
This could lead to the wrong library being chosen for non-Python
dependencies (e.g. libgfortran). To avoid this problem the library is
now passed with an absolute path.

**Additional changes**

- Remove unused code from `python_wrapper.py`
- Calculate includes/libs/libdirs/etc when needed to ensure paths are up
to date (and therefore absolute)