Ensure temporary arrays created for function calls are correctly declared in loops. (#1380)
When a function call takes an array as an argument and that array is the
result of an arithmetic operation a temporary variable is created. The
creation of new arrays is inserted before each line in a `CodeBlock`,
however despite the fact that `For.body` contains a `CodeBlock` it
didn't call `_visit_CodeBlock`. As a result the insertion of the line
was done before the `For` block instead of before the relevant line.

The problem is fixed by using `_visit_CodeBlock`. A test which
fails on the master branch has been added. Fixes #1379.