Rewrite bind C wrapper (#1388)
Previously this stage was conducted inside `ast.bind_c` however it
is not an AST object. The result was a new function which is then
wrapped in the C code. As the objects inside this function were not
linked (e.g. Variable to its sizes) there were lots of
`if language == 'fortran'` blocks in the cwrapper printer.

This PR introduces new objects for each of these concepts. As a result
the cwrapper can now be split into 2 parts. A printing part which just
prints, and a wrapping part which will be able to differentiate between
C and Fortran functions based on type in a similar way to what is done
in the rest of the code.

As part of this rewrite the handling of optionals was changed (fixes
#929) and strides were added (fixes #1117). Tests have been added for
the strides, the rest of the code is tested with the existing tests.
Documentation has also been added which must be increased when the
cwrapper is rewritten.

Some of this PR contains hacky fixes for the cwrapper. This is deemed
acceptable as the wrapper should be rewritten ASAP.