Rename `__pyccel__` and `__epyccel__` folders (#1451)
Change name of `__pyccel__` and `__epyccel__` folders to avoid
collisions during tests. The new name is the old name with an additional
suffix defined as the contents of the environment variable
`PYTEST_XDIST_WORKER` (when the variable is not present there is no
suffix). The tests now run around 10mins faster. Fixes #969.

Add a test for the `--out` argument.

**Bot fixes**
- Remove review handler from old bot
- Correct error leading to tests being triggered twice when PR is marked
as ready
- Correct error where labels are not removed when PR is made draft
- Correct documentation comparison folder
- Correct spelling relative directory
- Improve documentation