Bot fixes (#1460)
Add new workflows to the bot to react to reviews and to run the tests on
the `devel` branch. Fix a bug which stopped the bot from noticing when a
PR was ready (Codacy not identified). Add a mechanism to avoid running
tests if their dependencies haven't been changed. Ensure that all
necessary tests work without a pull request (for tests on the `devel`
branch).

---------

Co-authored-by: Mohammed Kaddani <ledevloppeurmohammedkaddani@gmail.com>