Bot fixes (#1466)
Try to fix additional bot problems:

1. Sort reviews before filtering them by author (hopefully will fix the
stage determination failing in #1280 and #1457). This should result in
only the most recent review being preserved.

2. Set flags if tests are already passing. If the tests were run before
marking the PR as draft the bot doesn't check the status and update the
flags as this was only done in the clean up. (Bug seen in #1457)

3. Fix a bug in the documentation test where it failed when a function
is removed from a class. (Bug seen in #1457)